BUGZID: 9829
Carambola : Unit tests for wvatomicfile


git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/carambola@9305 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ streams/wvatomicfile.cc
@@ -34,7 +34,7 @@ bool WvAtomicFile::open(WvStringParm filename, int mode, int create_mode)
     { 
         struct stat old_file;
         int fexists = lstat(atomic_file, &old_file);
-        if (!S_ISREG(old_file.st_mode))
+        if (!fexists && !S_ISREG(old_file.st_mode))
         {
             close();
             unlink(tmp_file);
@@ -45,6 +45,8 @@ bool WvAtomicFile::open(WvStringParm filename, int mode, int create_mode)
         tmp_file = WvString(tmptemplate(nitix));
         tmpfd = mkstemp(tmp_file.edit());
         fcntl(tmpfd, F_SETFL, mode);
+
+
 
         if (!WvFile::open(tmpfd))
             return false;
